ToDo Juan merge

+++ src/main/java/simulator/model/MapInfo.java
@@ -3,8 +3,8 @@ package simulator.model;
 public interface MapInfo extends JSONable {
     public int getCols();
     public int getRows();
-    public int getWidth();
-    public int getHeight();
-    public int getRegionWidth();
-    public int getRegionHeight();
+    public double getWidth();
+    public double getHeight();
+    public double getRegionWidth();
+    public double getRegionHeight();
 }
